Split devtools hooks
